refactor: 사용자 쿠폰 발급 처리 개별 트랜잭션으로 분리
