+++ src/main/java/couponToy/CouponToyProject/CouponIssue/service/IssueCouponService.java
@@ -13,11 +13,9 @@ import couponToy.CouponToyProject.global.exception.CouponNotFoundException;
 import couponToy.CouponToyProject.global.exception.MemberNotFoundException;
 import couponToy.CouponToyProject.global.security.MemberDetails;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class IssueCouponService {
@@ -40,25 +38,6 @@ public class IssueCouponService {
         IssueCoupon issueCoupon = issueCouponRepository.save(issueCouponRequest.toEntity(member, coupon));
 
         return IssueCouponResponse.fromEntity(issueCoupon);
-    }
-
-
-    @Transactional
-    public void issueFromQueue(Long couponId, Long userId) {
-        // 1. 쿠폰을 비관적 락으로 조회
-        Coupon coupon = couponRepository.findByCouponIdForUpdate(couponId)
-                .orElseThrow(() -> new CouponNotFoundException(ErrorCode.NOT_FOUND_COUPON));
-
-        // 2. 발급 수량 초과 여부 확인 & 발급 수량 증가
-        coupon.increaseIssueAmount();
-
-        // 3. 발급 이력 저장
-        IssueCoupon issueCoupon = IssueCoupon.builder()
-                .memberId(userId)
-                .couponId(couponId)
-                .build();
-
-        issueCouponRepository.save(issueCoupon);
     }
 }
